+++ src/unittest/CMakeLists.txt
@@ -1,0 +1,18 @@
+set (UNITTEST_SRCS
+	${CMAKE_CURRENT_SOURCE_DIR}/test.cpp
+	${CMAKE_CURRENT_SOURCE_DIR}/test_collision.cpp
+	${CMAKE_CURRENT_SOURCE_DIR}/test_compression.cpp
+	${CMAKE_CURRENT_SOURCE_DIR}/test_connection.cpp
+	${CMAKE_CURRENT_SOURCE_DIR}/test_filepath.cpp
+	${CMAKE_CURRENT_SOURCE_DIR}/test_inventory.cpp
+	${CMAKE_CURRENT_SOURCE_DIR}/test_mapnode.cpp
+	${CMAKE_CURRENT_SOURCE_DIR}/test_nodedef.cpp
+	${CMAKE_CURRENT_SOURCE_DIR}/test_noise.cpp
+	${CMAKE_CURRENT_SOURCE_DIR}/test_profiler.cpp
+	${CMAKE_CURRENT_SOURCE_DIR}/test_serialization.cpp
+	${CMAKE_CURRENT_SOURCE_DIR}/test_settings.cpp
+	${CMAKE_CURRENT_SOURCE_DIR}/test_socket.cpp
+	${CMAKE_CURRENT_SOURCE_DIR}/test_utilities.cpp
+	${CMAKE_CURRENT_SOURCE_DIR}/test_voxelalgorithms.cpp
+	${CMAKE_CURRENT_SOURCE_DIR}/test_voxelmanipulator.cpp
+	PARENT_SCOPE)
